+++ bin/server-config.php
@@ -24,7 +24,7 @@ try {
 
     $vpnConfigDir = sprintf('%s/openvpn-config', $baseDir);
     $serverClient = new ServerClient(
-        new CurlHttpClient([$config->requireString('apiUser'), $config->requireString('apiPass')]),
+        new CurlHttpClient($config->requireString('apiUser'), $config->requireString('apiPass')),
         $config->requireString('apiUri')
     );