+++ libexec/client-disconnect.php
@@ -45,7 +45,7 @@ try {
     );
 
     $serverClient = new ServerClient(
-        new CurlHttpClient([$config->requireString('apiUser'), $config->requireString('apiPass')]),
+        new CurlHttpClient($config->requireString('apiUser'), $config->requireString('apiPass')),
         $config->requireString('apiUri')
     );